+++ src/components/general-widgets/OwnerLogin/utils/getForgotPasswordFormMarkup.spec.js
@@ -1,38 +1,52 @@
-import React from 'react';
-
-import { Form } from 'collections/Form';
-import { Modal } from 'elements/Modal';
-import { TextInput } from 'inputs/TextInput';
+import { mount } from 'enzyme';
 
 import { getForgotPasswordFormMarkup } from './getForgotPasswordFormMarkup';
 
-describe('getForgotPasswordFormMarkup', () => {
-  it('should return the right markup ', () => {
-    const someFunction = () => {};
-    const forgotPasswordSubmitButtonText = 'A';
-    const forgotPasswordEmailInputLabel = 'B';
-    const forgotPasswordHeadingText = 'C';
-    const forgotPasswordModalTriggerText = 'D';
-    const forgotPasswordValidation = {};
-    const actual = getForgotPasswordFormMarkup(
+const someFunction = () => {};
+const forgotPasswordSubmitButtonText = 'A';
+const forgotPasswordEmailInputLabel = 'B';
+const forgotPasswordHeadingText = 'C';
+const forgotPasswordModalTriggerText = 'D';
+const forgotPasswordValidation = {};
+
+const getMarkup = ({ errorMessage = '', successMessage = '' }) =>
+  mount(
+    getForgotPasswordFormMarkup(
       someFunction,
       forgotPasswordSubmitButtonText,
       forgotPasswordEmailInputLabel,
       forgotPasswordHeadingText,
-      forgotPasswordModalTriggerText
-    );
+      forgotPasswordModalTriggerText,
+      forgotPasswordValidation,
+      errorMessage,
+      successMessage
+    )
+  );
 
-    expect(actual).toEqual(
-      <Modal trigger={<span>{forgotPasswordModalTriggerText}</span>}>
-        <Form
-          headingText={forgotPasswordHeadingText}
-          onSubmit={someFunction}
-          submitButtonText={forgotPasswordSubmitButtonText}
-          validation={forgotPasswordValidation}
-        >
-          <TextInput label={forgotPasswordEmailInputLabel} name="email" />
-        </Form>
-      </Modal>
-    );
+describe('getForgotPasswordFormMarkup', () => {
+  it('should return the right structure ', () => {
+    const wrapper = getMarkup({});
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  describe('if `forgotPasswordSuccessMessage` is passed', () => {
+    it('should render the success message above the reset button', () => {
+      const wrapper = getMarkup({
+        successMessage: 'This is a successful message',
+      });
+
+      expect(wrapper).toMatchSnapshot();
+    });
+  });
+
+  describe('if `forgotPasswordErrorMessage` is passed', () => {
+    it('should render the error message above the reset button', () => {
+      const wrapper = getMarkup({
+        errorMessage: 'This is an error message',
+      });
+
+      expect(wrapper).toMatchSnapshot();
+    });
   });
 });