+++ src/components/general-widgets/OwnerSignUp/component.js
@@ -17,17 +17,21 @@ import { TextInput } from 'inputs/TextInput';
 // eslint-disable-next-line jsdoc/require-jsdoc
 export const Component = ({
   emailInputLabel,
+  errorMessage,
   firstNameInputLabel,
   headingText,
   lastNameInputLabel,
   onSubmit,
   submitButtonText,
+  successMessage,
   validation,
 }) => (
   <Form
+    errorMessage={errorMessage}
     headingText={headingText}
     onSubmit={onSubmit}
     submitButtonText={submitButtonText}
+    successMessage={successMessage}
     validation={validation}
   >
     <TextInput label={firstNameInputLabel} name="firstName" />
@@ -40,17 +44,21 @@ Component.displayName = 'OwnerSignUp';
 
 Component.defaultProps = {
   emailInputLabel: EMAIL,
+  errorMessage: '',
   firstNameInputLabel: FIRST_NAME,
   headingText: OWNER_SIGNUP,
   lastNameInputLabel: LAST_NAME,
   onSubmit: Function.prototype,
   submitButtonText: SIGN_UP,
+  successMessage: '',
   validation: {},
 };
 
 Component.propTypes = {
   /** The label for the email input. */
   emailInputLabel: PropTypes.string,
+  /** The message to display when the form has an error. */
+  errorMessage: PropTypes.string,
   /** The label for the first name input. */
   firstNameInputLabel: PropTypes.string,
   /** The text to display as a heading at the top of the widget. */
@@ -63,6 +71,8 @@ Component.propTypes = {
   onSubmit: PropTypes.func,
   /** The text to display on the submit button. */
   submitButtonText: PropTypes.string,
+  /** The message to display when the form is successful. */
+  successMessage: PropTypes.string,
   /** Settings for validating inputs. Each value should match [the shape documented in `Form`](https://lodgify.github.io/lodgify-ui/#/Collections/Form) */
   validation: PropTypes.shape({
     email: PropTypes.object,