+++ src/components/media/ResponsiveImage/utils/getLowestCommonDenominator.spec.js
@@ -1,0 +1,9 @@
+import { getLowestCommonDenominator } from './getLowestCommonDenominator';
+
+describe('`getLowestCommonDenominator`', () => {
+  it('should return the right value', () => {
+    const actual = getLowestCommonDenominator(1, 4);
+
+    expect(actual).toMatchSnapshot();
+  });
+});